[FIX] tests for TYPO3 13 @ 2024.07.09
Relates: #3995
